+++ 2023/src/day/day07.cpp
@@ -1,4 +1,5 @@
 #include <algorithm>
+#include <cstddef>
 #include <iostream>
 #include <string>
 #include <unordered_map>
@@ -131,7 +132,7 @@ void aoc::day07(std::vector<std::string> input) {
   });
 
   int answer1 = 0;
-  for (int i = 0; i < cards.size(); i++) {
+  for (size_t i = 0; i < cards.size(); i++) {
     answer1 += (i + 1) * cards[i].bid;
   }
 
@@ -144,7 +145,7 @@ void aoc::day07(std::vector<std::string> input) {
   });
 
   int answer2 = 0;
-  for (int i = 0; i < cards.size(); i++) {
+  for (size_t i = 0; i < cards.size(); i++) {
     answer2 += (i + 1) * cards[i].bid;
   }
